ft_bzero: Use size_t instead of unsigned int

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: daniema3 <daniema3@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/16 19:32:22 by daniema3          #+#    #+#             */
-/*   Updated: 2024/09/18 18:54:25 by daniema3         ###   ########.fr       */
+/*   Updated: 2024/09/18 21:18:20 by daniema3         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void	*str, size_t n)
 {
-	unsigned int	i;
+	size_t			i;
 	unsigned char	*casted_str;
 
 	i = 0;
